+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Enhancements ###
 
 - It is now possible to override and specify new parameters for OAuth2 authentication URL by using "oauth2." prefix in security_details key name.
+- Security details values can be loaded from environment variables.
 
 ## 0.61 (2017-06-22) ##